+++ src/compiler/typescript_to_rust/azle_generate/src/candid_type/type_alias/mod.rs
@@ -5,8 +5,8 @@ use crate::{traits::GetName, ts_ast::SourceMapped};
 
 impl SourceMapped<'_, TsTypeAliasDecl> {
     pub fn to_type_alias(&self) -> Option<TypeAlias> {
-        self.process_ts_type_ref("Alias", |azle_type_ref| {
-            let aliased_type = azle_type_ref.get_ts_type().to_candid_type();
+        self.process_ts_type_ref("Alias", |type_ref| {
+            let aliased_type = type_ref.get_ts_type().to_candid_type();
 
             let type_params = self.get_type_params();
 
@@ -27,8 +27,8 @@ impl SourceMapped<'_, TsTypeAliasDecl> {
                 swc_ecma_ast::TsEntityName::TsQualifiedName(_) => None,
                 swc_ecma_ast::TsEntityName::Ident(ident) => {
                     if ident.get_name() == type_name {
-                        let azle_type_ref = SourceMapped::new(ts_type_ref, self.source_map);
-                        Some(handler(azle_type_ref))
+                        let type_ref = SourceMapped::new(ts_type_ref, self.source_map);
+                        Some(handler(type_ref))
                     } else {
                         None
                     }
