+++ src/compiler/typescript_to_rust/azle_generate/src/canister_method/query_and_update/shared/build_canister_methods.rs
@@ -19,9 +19,9 @@ impl BuildCanisterMethods for Vec<AzleProgram> {
         &self,
         canister_method_type: CanisterMethodType,
     ) -> Vec<CanisterMethod> {
-        let azle_fnc_decls = self.get_annotated_fn_decls_of_type(canister_method_type.clone());
+        let fnc_decls = self.get_annotated_fn_decls_of_type(canister_method_type.clone());
 
-        azle_fnc_decls.iter().fold(vec![], |acc, fn_decl| {
+        fnc_decls.iter().fold(vec![], |acc, fn_decl| {
             let canister_method_node = fn_decl.build_canister_method_node(&canister_method_type);
             vec![acc, vec![canister_method_node]].concat()
         })
