+++ src/compiler/typescript_to_rust/azle_generate/src/ts_ast/ts_type/to_candid_type.rs
@@ -9,12 +9,9 @@ impl SourceMapped<'_, TsType> {
     pub fn to_candid_type(&self) -> CandidType {
         match self.deref() {
             TsType::TsKeywordType(x) => SourceMapped::new(x, self.source_map).to_candid_type(),
-            TsType::TsFnOrConstructorType(x) => match x {
-                TsFnOrConstructorType::TsFnType(x) => {
-                    SourceMapped::new(x, self.source_map).to_candid_type()
-                }
-                TsFnOrConstructorType::TsConstructorType(_) => todo!(),
-            },
+            TsType::TsFnOrConstructorType(TsFnOrConstructorType::TsFnType(x)) => {
+                SourceMapped::new(x, self.source_map).to_candid_type()
+            }
             TsType::TsTypeRef(x) => SourceMapped::new(x, self.source_map).to_candid_type(),
             TsType::TsTypeLit(_) => {
                 let origin = self.get_origin();
@@ -32,7 +29,17 @@ impl SourceMapped<'_, TsType> {
 
                 panic!("Unexpected TsTupleType\n     at {}:{}:{}\n\nHelp: Try wrapping this with Tuple", origin, line_number, column_number)
             }
-            _ => todo!(),
+            _ => {
+                let origin = self.get_origin();
+                let line_number = self.get_line_number();
+                let range = self.get_range();
+                let column_number = range.0 + 1;
+
+                panic!(
+                    "Unexpected TsType\n     at {}:{}:{}\n\nHelp: Try removing this type",
+                    origin, line_number, column_number
+                )
+            }
         }
     }
 }
