+++ src/compiler/typescript_to_rust/azle_generate/src/ts_ast/ts_type/ts_fn_or_constructor_type/mod.rs
@@ -1,5 +1,2 @@
-mod errors;
 mod get_span;
 mod ts_fn_type;
-
-// TODO: Remove all occurrences of `from_ts_fn_or_constructor_type`
